% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{hrPool}
\alias{hrPool}
\title{Empirical UMP beta pooled p-value}
\usage{
hrPool(p, w = 1, nsim = 1e+05)
}
\arguments{
\item{p}{numeric vector of p-values between 0 and 1}

\item{w}{numeric value between 0 and 1}

\item{nsim}{integer, the number of simulated null cases generated}
}
\value{
A numeric value between 0 and 1.
}
\description{
Uses simulation under the null to approximate the UMP
pooled p-value for a restricted beta family.
}
\details{
To test the null hypotheses that all p-values are uniform
against a restricted beta family 0 < a <= 1 <= b, the most
powerful pooled p-value linearly combines upper and lower tail
probabilities of the chi-squared distribution with two degrees
of freedom with weights w and (1 - w) where w = (1 - a)/(b - a).

This function computes the statistic given by this combination
for a collection of p-values, and then simulates a specified
number of null cases to give an empirical pooled p-value.
}
\examples{
p <- c(0.1, 0.5, 0.9)
hrPool(p, 0.2)
hrPool(p, 0.5)
hrPool(p, 0.9)
}
\author{
Chris Salahub
}
